Increase modularity of packages and implement interfaces
